Chane ENV to DBG
ENV refers to all cgi-bin scripts which we use for debugging. This
was ok when the only script was the env script but now that we have
several scripts, the name is being changed to DBG.

+++ test/features/local/targets.mk
@@ -26,7 +26,7 @@ $(TESTS):
 	behave $$FLAGS $@.feature
 
 down:
-	docker-compose down
+	-test -f .up && docker-compose down
 	-rm .up
 
 clean:
